+++ frontend/src/App.js
@@ -1,15 +1,32 @@
 import React from "react";
-import { Route, Switch } from "react-router-dom";
-// import SignupFormPage from "./components/SignupFormPage";
-// // import LoginFormPage from "./components/LoginFormPage";
-// import Navigation from "./components/Navigation";
-import Navbar from "./components/Navbar";
+import { useSelector } from "react-redux";
+import UserPage from "./components/Navbar/UserPage";
+import WelcomePage from "./components/Navbar/WelcomePage";
 
 function App() {
+  const sessionUser = useSelector(state => state.session.user)
+
+  let sessionLinks;
+
+  if (sessionUser) {
+    sessionLinks = (
+      <>
+        <UserPage user={sessionUser} />
+      </>
+    );
+  } else {
+    sessionLinks = (
+      <>
+        <WelcomePage />
+      </>
+    )
+
+  }
   return (
     <>
-      <Navbar/>
+      {sessionLinks}
     </>
+  )
     // <>
     //   <Navigation />
     //     <Switch>
@@ -21,7 +38,6 @@ function App() {
     //       </Route>
     //     </Switch>
     // </>
-  );
 }
 
 export default App;
